btf: add CO-RE enum relocations
Support relocating enum values. The relocation happens by comparing the
(essential) name of the local enum value with the name of the target
enum. If the names match the target value is used. If no match is found,
the relocation returns 0.

This also fixes a bug where typeID relocations used the target ID twice.